Combine classes into one command
