+++ src/commands/classes/classes.ts
@@ -3,14 +3,14 @@ import Command from '../../Command';
 import SubCommand from '../../SubCommand';
 import CommandOption from '../../CommandOption';
 import { CommandOptionType } from '../../Discord';
-import { dbClass, classExists, createClass } from './_common';
+import { dbClass, classExists, createClass, getRole } from './_common';
 
-const newclass = new Command({
-  name: 'newclass',
-  description: 'Create a new class role assignable by users',
+const classes = new Command({
+  name: 'classes',
+  description: 'Manage server classes',
   options: [
     new SubCommand({
-      name: 'class',
+      name: 'create',
       displayName: 'New Class',
       description: 'Creates a new class role',
       requiredPermissions: [Discord.Permission.MANAGE_ROLES],
@@ -52,12 +52,47 @@ const newclass = new Command({
 
           // Notify of successful class creation
           return ctx.respondWithMessage(
-            `Created class **${name}**`,
+            `Created role for class **${name}**`,
           );
+        }
+      },
+    }),
+    new SubCommand({
+      name: 'join',
+      displayName: 'Join Class',
+      description: 'Joins a class',
+      requiredPermissions: [Discord.Permission.MANAGE_ROLES],
+      options: [
+        new CommandOption({
+          name: 'name',
+          description: 'The class name/code',
+          required: true,
+          type: CommandOptionType.STRING,
+        }),
+      ],
+      handler: async ctx => {
+        const guildId = ctx.mustGetGuildId();
+        const name = ctx.getArgument<string>('name')?.trim() as string;
+        if (!await classExists(ctx, guildId, name)) {
+          ctx.respondWithError(`That class does not exist`);
+        } else {
+            // Get the class role
+            const classRole = await getRole(ctx, guildId, name);
+            const roleId = classRole.id;
+            const userId = ctx.interaction.member?.user?.id;
+            
+            if (userId) {
+                await ctx.api.addRoleToMember(guildId, userId, roleId);
+
+                // Notify of successful class creation
+                return ctx.respondWithMessage(
+                `Joined class **${name}**`,
+                );
+            }
         }
       },
     }),
   ],
 });
 
-export default newclass;
+export default classes;
