+++ src/commands/classes/index.ts
@@ -1,2 +1,1 @@
-export { default as newclass } from './newclass';
-export { default as joinclass } from './joinclass';
+export { default as classes } from './classes';
